reduced interaction_matrix dtype to uint32

+++ recommendation_module.py
@@ -26,9 +26,9 @@ class Item2Item:
         # select columns to work on
         df = data.loc[:, [user_col, item_col]].drop_duplicates(subset=[user_col, item_col])
 
-        interaction_matrix = df.merge(df, on=user_col, how="outer").groupby([item_col + "_x", item_col + "_y"], observed=True).size().unstack().fillna(0)
+        interaction_matrix = df.merge(df, on=user_col, how="outer").groupby([item_col + "_x", item_col + "_y"], observed=True).size().unstack().fillna(0).astype("uint32")
         
-        self.item2item_frequency = self.item2item_frequency.add(interaction_matrix, fill_value=0).fillna(0)
+        self.item2item_frequency = self.item2item_frequency.add(interaction_matrix, fill_value=0).fillna(0).astype("uint32")
         assert (self.item2item_frequency.index == self.item2item_frequency.columns).all()
 
     def estimate_scores(self):
